use shared interface to make sure new sql driver doesn't generate tipset cache
